User registration

+++ base/views.py
@@ -1,17 +1,25 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
 
+
 def login_page(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('base:home')
 
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try: 
@@ -29,13 +37,32 @@ def login_page(request):
             messages.error(request, 'Username OR password does not exist.')
 
 
-    context= {}
-    return render(request, 'base/login_resgister.html', context)
+    context= {'page': page}
+    return render(request, 'base/login_register.html', context)
+
 
 def logout_user(request):
     logout(request)
     return redirect('base:home')
 
+
+def register_page(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(data=request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('base:home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+    
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request):
@@ -60,6 +87,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='base:login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -72,10 +101,14 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='base:login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     # Initial request, prefill form with the current entry.
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         # POST data submitted; process data
@@ -88,8 +121,12 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='base:login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == "POST":
         room.delete()
